Added requestAudioFocus, removeAudioFocus methods
Added the request and removeAudiofocus method to check if there are other
media files playing in the background.these prevent the app from playing
a mediaFile while other media files are playing in the background.

Also, added a onStartCommand to tell the MediaPlayerService which methods
to run to start the service.